Added install / uninstall file
Re-organized for easy deployment should be simple to add an update
function to gpsy-conf.sh so that new screenrc profiles can be added /
removed.

the only issue with that is that we'll need to ensure that we don't
clobber the existing gpsy because it could have user specific stuff.

maybe consider rewriting that so that user specific stuff is totally
separate.